+++ app/javascript/vue/tasks/citations/otus/store/mutations/setSourceSelected.js
@@ -1,3 +1,0 @@
-export default function (state, newSource) {
-  state.selected.source = newSource
-}