+++ app/javascript/vue/tasks/collection_objects/browse/main.js
@@ -1,9 +1,17 @@
 import { createApp } from 'vue'
 import { newStore } from './store/store.js'
 import App from './App.vue'
+import HelpSystem from '@/plugins/help/help'
+import en from './help/en'
 
 function initApp(element) {
   const app = createApp(App)
+ 
+  app.use(HelpSystem, {
+    languages: {
+      en
+    }
+  })
 
   app.use(newStore())
   app.mount(element)