+++ app/javascript/vue/tasks/data_attributes/multi_update/adapters/makeDataAttribute.js
@@ -3,7 +3,7 @@ export function makeDataAttribute(da) {
   return {
     id: da.id,
     uuid: randomUUID(),
-    value: da.value,
+    value: da.value || '',
     predicateId: da.controlled_vocabulary_term_id,
     objectId: da.attribute_subject_id,
     objectType: da.attribute_subject_type,