+++ app/javascript/vue/tasks/field_occurrences/browse/constants/collectingEvents.js
@@ -1,0 +1,26 @@
+export const PRIORITIZE_ATTRIBUTES = [
+  'id',
+  'verbatim_label',
+  'verbatim_collectors',
+  'verbatim_method',
+  'verbatim_habitat',
+  'verbatim_locality',
+  'verbatim_latitude',
+  'verbatim_longitude',
+  'verbatim_date'
+]
+export const HIDE_ATTRIBUTES = [
+  'global_id',
+  'geographic_area_id',
+  'identifiers',
+  'url_for',
+  'project_id',
+  'object_url',
+  'object_label',
+  'object_tag',
+  'cached',
+  'created_by_id',
+  'updated_by_id',
+  'base_class',
+  'md5_of_verbatim_label'
+]