+++ app/models/field_occurrence.rb
@@ -21,29 +21,28 @@ class FieldOccurrence < ApplicationRecord
   include Housekeeping
 
   include Shared::Citations
+  include Shared::Confidences
   include Shared::DataAttributes
+  include Shared::Depictions
+  include Shared::HasPapertrail
   include Shared::Identifiers
   include Shared::Notes
-  include Shared::Tags
-  include Shared::Depictions
-
-  include Shared::OriginRelationship
-  include Shared::Confidences
-  include Shared::ProtocolRelationships
-  include Shared::HasPapertrail
   include Shared::Observations
+  include Shared::OriginRelationship
+  include Shared::ProtocolRelationships
+  include Shared::Tags
   include Shared::IsData
   include Shared::QueryBatchUpdate
   include SoftValidation
 
   # At present must be before BiologicalExtensions
   include Shared::TaxonDeterminationRequired
-  # At present must be before IsDwcOccurence
-  include FieldOccurrence::DwcExtensions
-  include Shared::Taxonomy
-
   include Shared::BiologicalExtensions
-  include Shared::IsDwcOccurrence
+
+  include Shared::Taxonomy
+  include FieldOccurrence::DwcExtensions
+
+  GRAPH_ENTRY_POINTS = [:biological_associations, :taxon_determinations, :biocuration_classifications, :collecting_event]
 
   is_origin_for 'Specimen', 'Lot', 'Extract', 'AssertedDistribution', 'Sequence'
 
@@ -63,6 +62,7 @@ class FieldOccurrence < ApplicationRecord
   validate :check_that_either_total_or_ranged_lot_category_id_is_present
   validate :check_that_both_of_category_and_total_are_not_present
   validate :total_zero_when_absent
+  validate :total_positive_when_present
 
   accepts_nested_attributes_for :collecting_event, allow_destroy: true, reject_if: :reject_collecting_event
 
@@ -74,6 +74,10 @@ class FieldOccurrence < ApplicationRecord
 
   def total_zero_when_absent
     errors.add(:total, 'Must be zero when absent.') if (total != 0) && is_absent
+  end
+
+  def total_positive_when_present
+    errors.add(:total, 'Must be positive when not absent.') if !is_absent && total.present? && total <= 0
   end
 
   def check_that_both_of_category_and_total_are_not_present